% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS accident data for given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{numeric state identifier of state to be mapped}

\item{year}{for which the accident should be mapped}
}
\value{
None
}
\description{
\code{fars_map_state()} makes use of several helper functions to map FARS
accident data. Respective data for the given year are plotted within a map of
the state specified by the state number \code{state.num}. Functions throws an
error if state number is not existent in data. Function does not return
anything.
}
\examples{
\dontrun{fars_map_state(1, 2014)}

}
