% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Import FARS data from compressed csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Character string giving the path to the csv file to be read}
}
\value{
returns tidy tibble of accident data
}
\description{
\code{fars_read()} imports data of the Fatality Analysis Reporting System
(FARS). The function makes use of the delivered compressed csv files. The
data is converted to and outputted as a tibble. Function throws an error if
file does not exist.
}
\examples{
\dontrun{fars_read('accident_2013.csv.bz2')}
}
