% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{FARS summary by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{vector of years or single year for which data is summarized}
}
\value{
summary by year and month
}
\description{
\code{fars_summarize_years()} computes monthly summary statistics (counts of
accidents) for a given year or vector of \code{years}.
}
\examples{
\dontrun{fars_summarize_years(2013)}
\dontrun{fars_summarize_years(c(2013,2014))}

}
