% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Creating FARS filenames according to naming standard}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year for which the filename should be constructed}
}
\value{
filename character of filename
}
\description{
Helper function to stitch together FARS data filename for a given \code{year}.
}
\examples{
\dontrun{make_filename(2014)}
}
